Add checks to update script

+++ utils/update.sh
@@ -6,19 +6,34 @@ SOURCE="/home/doctormo/bazaar/inkscape-web"
 TARGET="/var/www/staging.inkscape.org/"
 EX="/var/www/staging.inkscape.org/keep-local.conf"
 
+SETTINGS="inkscape/settings.py"
+REQUIRES="utils/requirements.txt"
+
+PYTHON="./pythonenv/bin/python"
+PIP="./pythonenv/bin/pip"
+
 cd $SOURCE
 
-bzr pull
+bzr pull | grep Now\ on\ revision &> /dev/null
+if [[ $? == 0 ]]; then
+    echo "New revisions found, updating!"
 
-rsync -rtuv $TEST --delete --exclude-from="$EX" . $TARGET
+    rsync -rtuv $TEST --delete --exclude-from="$EX" . $TARGET
 
-cd $TARGET
+    cd $TARGET
 
-./pythonenv/bin/pip install -r utils/requirements.txt
+    if [ $REQUIRES -nt $PIP ]; then
+        $PIP install -r $REQUIRES
+    fi
 
-./pythonenv/bin/python inkscape/manage.py syncdb
-./pythonenv/bin/python inkscape/manage.py migrate
+    if [ $SETTINGS -nt "data/updated.last" ]; then
+        $PYTHON inkscape/manage.py syncdb
+        $PYTHON inkscape/manage.py migrate
+        touch "data/updated.last"
 
-# Restart server like this:
-touch utils/wsgi_load.py
-
+        # Restart server like this:
+        touch utils/wsgi_load.py
+    fi
+else
+    echo "No new revisions, no update."
+fi
